+++ condow_core/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] -  - 2021/12/11
+
+### Added
+
+- `value` const methods added to units `Mebi`, `Gibi`, etc for allowing them to be used to initialize constants
+- Documentation for async reader
+
+### Changes
+
+- Breaking: In meory clients are generic over the location. Default tyoe is `NoLocation`
+- Updated README.md
+
 ## [0.10.2] - 2021/10/22
 
 ### Added